+++ b2sdk/v1/account_info.py
@@ -8,6 +8,7 @@
 #
 ######################################################################
 import inspect
+from typing import Optional
 
 from b2sdk import _v2 as v2
 
@@ -58,6 +59,12 @@ class AbstractAccountInfo(OldAccountInfoMethods, v2.AbstractAccountInfo):
         :rtype: str
         """
         # Removed @abstractmethod decorators
+
+    def get_bucket_name_or_none_from_bucket_id(self, bucket_id: str) -> Optional[str]:
+        """
+        Look up the bucket name for the given bucket id.
+        """
+        # Removed @abstractmethod decorator
 
 
 class InMemoryAccountInfo(v2.InMemoryAccountInfo, AbstractAccountInfo):